move to type script in web extension sample
