+++ Samples/HelloWorld/Project/HelloWorld.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="DotNetJS" Version="0.3.8" />
+        <PackageReference Include="DotNetJS" Version="0.3.10" />
     </ItemGroup>
 
 </Project>
