+++ Samples/WebExtension/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "target": "es6",
+    "module": "commonjs",
+    "strict": true
+  },
+  "include": ["src/**/*.ts"]
+}
